CN 屏蔽拆成独立文件 cn-block.py，菜单加「卸载」
- 屏蔽中国域名/IP 的全部逻辑移到独立文件 cn-block.py，方便单独维护；
  主脚本只负责拉最新 cn-block.py 到 /etc/bgpeer 并子进程调用：
    菜单第 9 项 → python3 cn-block.py（交互）
    重装后     → python3 cn-block.py apply（仅已开启时重注入）
- cn-block.py 菜单新增「4 卸载」：清掉注入规则 + 删除状态文件，
  不想屏蔽直接卸载即可（之后重装也不会再自动注入）；原退出改为 5
- cn-block.py 也支持 apply/remove 命令行参数，可完全脱离主脚本单跑
